Refactor root template to its own file
